# docs/openapi/heimdall.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Heimdall API",
    "description": "A centralized identity management API with scope-based multi-tenancy.\n\nEvery identifier in a path is a **PublicId** (GUID), never an internal id. Every endpoint answers one of two envelopes — `DataOutput<T>` for a single resource, `PaginatedOutput<T>` for a listing.\n\nObtain a token from `POST /api/auth/login`, then authorize with it. A challenge token issued by a 2FA-gated login is rejected everywhere except `POST /api/auth/2fa/verify`.",
    "contact": {
      "name": "Artur Rios",
      "url": "https://github.com/artur-rios/heimdall-api"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/artur-rios/heimdall-api/blob/main/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Authenticates a person by email and password and returns a token (UC-11, FR-AU-01…07). A\n`User` also sends the `PublicId` of their scope; a `ScopeAdmin` or\n`SystemAdmin` sends credentials only. Open to anonymous callers — this is where a\ncaller gets the token every other endpoint requires. Every rejection (AF-11a…AF-11e)\nanswers 401 alike, so the endpoint cannot be used to enumerate accounts.",
        "description": "**Anonymous** — no bearer token required.",
        "operationId": "Auth_Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoginCommandOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginCommandOutputDataOutput"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
    "/api/auth/password-recovery": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Requests a password reset link (UC-12, FR-PR-01/02). A `User` also sends the\n`PublicId` of their scope; a `ScopeAdmin` or `SystemAdmin` sends the email\nalone. Open to anonymous callers — someone who has lost their password cannot hold a\ntoken.",
        "description": "Answers 200 with the same message whether or not the address belongs to anyone (AF-12a),\nso the endpoint cannot be used to enumerate accounts. The only rejection is a malformed\nrequest (400, NFR-10), which says nothing about who is registered.\n\n**Anonymous** — no bearer token required.",
        "operationId": "Auth_PasswordRecovery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordRecoveryCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordRecoveryCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordRecoveryCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordRecoveryCommandOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordRecoveryCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordRecoveryCommandOutputDataOutput"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
    "/api/auth/password-reset": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sets a new password from the reset token mailed by UC-12 (UC-13, FR-PR-03/04). Open to\nanonymous callers for the same reason: the token is the only credential someone who has\nlost their password can present.",
        "description": "Unlike the two endpoints above, each rejection is named — unknown (AF-13c), expired\n(AF-13a), and spent (AF-13b) tokens all answer 400 with their own message, as does a\nmalformed request (AF-13d). Nothing is disclosed by the distinction: the token identifies\nno account to a caller who does not already hold it.\n\n**Anonymous** — no bearer token required.",
        "operationId": "Auth_ResetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordCommandOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordCommandOutputDataOutput"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Confirms an email address from the verification token mailed at person creation (UC-14,\nFR-EV-03). Open to anonymous callers: the person reaches this from a link in their mail\nclient, where they hold no bearer token — and the point of the link is that they have not\nproved anything yet.",
        "description": "Each rejection is named, as UC-13's are — unknown (AF-14c), expired (AF-14a), and spent\n(AF-14b) tokens all answer 400 with their own message, as does a request carrying no token\nat all. An address that was already verified answers 200: UC-14 defines no alternative flow\nfor it, and the link did what it promised.\n\n**Anonymous** — no bearer token required.",
        "operationId": "Auth_VerifyEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailCommandOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailCommandOutputDataOutput"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
    "/api/auth/resend-verification": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Retires the caller's outstanding verification links and mails a fresh one (UC-15,\nFR-EV-04), for someone whose first link expired, was lost, or never arrived. The one\nauthenticated endpoint on this controller — and the reason it takes no request body: the\nperson is read from the bearer token, so a caller can only ever ask for their own link.",
        "description": "No `RoleRequirement`: the authorization matrix grants email verification to all three\nroles and withholds it from anonymous callers, which is exactly what authentication alone\nenforces. An address that is already verified answers 400 (AF-15a) — unlike UC-14's\nidempotent success, since a link mailed to a verified address could do nothing when clicked.\nA token naming a person who no longer exists answers 404: it was validated, but there is no\naddress left to send to.\n\n**Any authenticated caller** — the handler decides who may act, so no role is required at the door.",
        "operationId": "Auth_ResendVerification",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResendVerificationEmailCommandOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResendVerificationEmailCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResendVerificationEmailCommandOutputDataOutput"
                }
              }
            }
          },
          "401": {
            "description": "The token is missing, malformed, expired, or is a 2FA challenge token."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/auth/2fa": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "**Any authenticated caller** — the handler decides who may act, so no role is required at the door.",
        "operationId": "Auth_GetTwoFactorStatus",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorStatusOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorStatusOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorStatusOutputDataOutput"
                }
              }
            }
          },
          "401": {
            "description": "The token is missing, malformed, expired, or is a 2FA challenge token."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginCommand": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "scopeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LoginCommandOutputDataOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/LoginCommandOutput"
          }
        },
        "additionalProperties": false
      },
...
      "ResetPasswordCommand": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste the token returned by POST /api/auth/login. The 'Bearer ' prefix is added for you.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}